Rework for : - batch class generic abstraction - reorganise code into coherent namespaces - rework composer's  psr4 loader to handle those new namespaces
